Add PyCommand 'findinstruction'
Add PyCommand 'findinstruction' for immunity debugger that searches for
a specified instruction in the open process's address space and returns
the address of matches, if they lie in an executable section.

This was written for Immunity Debugger 1.85, whose immlib has some
differences then the version used in the book, so the sample code does
not match.