Add ftp fuzzer that uses Sulley
Add the ftp skeleton as well as the starting script to fuzz and ftp
server.

Setup:
Windows XP (VM) running WarFTPD version 1.65 on port 1.65
Sulley on the VM running the process monitor on port 26002
Sulley on the host machine running the network monitor on port 26001
VirtualBox set to bridge the network. Guest got 192.168.1.107.

Sulley doesn't appear to fuzz deterministically. I went through all
'user' cases without failure, and then after restarting got a crash
on the 49th iteration.

The process monitor didn't appear to work: The pcap files are written,
but are only a few bytes long and couldn't be parsed by Wireshark.