Add ioctl_dump and simple ioctl fuzzer
Add ioctl_dump which is a PyCommand for Immunity that tries to find
device strings in a debugged driver as well as the IOCTL command values
that it can accept. It pickles these and saves them to disk.

my_ioctl_fuzzer is a simple script that reads a pickle to get a list of
device names and ioctl commands. It checks which of the device names
can be accessed and then fuzzes random data to random devices with
random ioctl commands.